+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **[Breaking]** Drop old-style `node` target.
 - **[Feature]** Add `mocha` target to run tests in `spec.ts` files.
 - **[Feature]** Add `node` target to build and run top-level Node applications.
+- **[Feature]** Provide `generateNodeTasks`, `generateLibTasks` and `generateMochaTasks` functions.
+  They create the tasks but do not register them. 
 - **[Fix]** Run `clean` before `dist`, if defined.
 - **[Fix]** Run `dist` before `publish`.